[opentitanlib] Add `--load-memory` option to `InitializeTest`
Now, even without going through OpenTitanTool separately, you can load
memories from just a host test binary by using e.g.
```
--load-memory ROM={VMEM} --load-memory OTP={VMEM}
```

There is some small refactoring in this case to reduce the duplication
with the OpenTitanTool command logic.
