+++ sw/host/opentitanlib/BUILD
@@ -147,6 +147,7 @@ rust_library(
         "src/test_utils/crashdump.rs",
         "src/test_utils/e2e_command.rs",
         "src/test_utils/epmp.rs",
+        "src/test_utils/fpga_backdoor.rs",
         "src/test_utils/gpio.rs",
         "src/test_utils/gpio_monitor.rs",
         "src/test_utils/i2c_target.rs",
